+++ src/LanguageServer/Impl/Resources.resx
@@ -132,6 +132,9 @@
   <data name="Error_InvalidCachePath" xml:space="preserve">
     <value>Specified cache folder does not exist. Switching to default.</value>
   </data>
+  <data name="ImportLocally" xml:space="preserve">
+    <value>import locally '{0}'</value>
+  </data>
   <data name="InitializingForGenericInterpreter" xml:space="preserve">
     <value>Initializing for generic interpreter</value>
   </data>
